requirements.txt
pytest==8.0.0
pytest-html==4.1.0
pytest-xdist==3.5.0
pytest-rerunfailures==13.0

pytest-playwright==0.5.2
playwright==1.57.0

allure-pytest==2.15.2
python-dotenv==1.0.1
requests==2.31.0

allure-python-commons